test: isolate GPT-OSS checkpoint parity
